T: rewrite proc macro tests with `WithProcMacroRustProjectDescriptor`
